refactor: change the file structure
