+++ week8/algoorgoal/src/components/common/OrderBar/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import useCartStore from '../../store/cartStore';
+import useCartStore from '../../../store/cartStore';
 import OrderAmountContainer from './OrderAmountContainer';
-import OrderButton from '../OrderButton';
-import Row from '../common/Row';
+import OrderButton from '../../OrderButton';
+import Row from '../Row';
 
 export default function OrderBar() {
   const menu = useCartStore((state) => state.menu);
