+++ week8/algoorgoal/src/stories/BackButton.stories.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react';
 
-import BackButton from '../components/NavigationRow/BackButton';
+import BackButton from '../components/common/NavigationRow/BackButton';
 
 const meta = {
   title: 'week8/BackButton',
